Try to make sure log file is world-writable
This is a hopefully-temporary hack to deal with the fact that our log
file is currently owned by whatever user ran bazels3cache. The problem
is that if one user logs into a machine and launches it, then the log
file will be owned by that user; later, if a different user logs in and
tries to launch bazels3cache, it wouldn't be able to write to the log.

As a workaround, when we create the log file, we make sure it is
world-writable.