Change default number of errors before pausing S3 access
The previous default for how many S3 errors we would allow before
pausing access to S3 was ten errors. This changes the default to three
errors. Three consecutive errors indicates a pretty high likelihood that
the user is offline; and lowering the default may speed up the user's
Bazel compilation, by spending less time trying to connect to the
network.