Remove call to process.stdin.end()
The daemon was calling stdin.end(), and I already had a comment
indicating that I wasn't sure it was necessary to do that. Not only is
it not necessary, but it turns out that for some reason, it was causing
bazels3cache to exit immediately when run from inside a Dockerfile.