Async upload
By default, uploads are now asynchronous. In other words, when Bazel
does a PUT to upload a file, bazels3cache saves the data locally (on
disk), and then immediately sends a success reply back to Bazel. Then it
starts the upload to the remote cache.  This can make Bazel builds much
faster, because Bazel does not need to wait for the upload to finish.