This fixes some cases where bazel hangs forever.
If the connection/socket bazel has opened with the remote cache
closes unexpectedly (without a response) then the bazel build
would hang forever. I replicated the effects we had observed here
https://app.asana.com/0/41418719780575/639438010491097/f
by setting a very small socket timeout. I fixed it on the bazels3cache
side by returning a 404 when the socket times out (instead of just closing
the socket) and made the socket timeout configurable and set it to
10 minutes instead of the (2-minute default).