Allow specifying maxSockets to s3 connection
By default s3 limits to 50 concurrent connections.[1] For builds with many concurrent
jobs this may not be enough, so here we introduce a setting allowing this limit to be
increased (or decreased). Note that the s3 client considers "Infinity" to mean 50, so
you will wish to choose a large, finite, number.

[1]  https://docs.aws.amazon.com/sdk-for-javascript/v2/developer-guide/node-configuring-maxsockets.html

Test Plan:
Ran locally with and without the setting, and observed requests being handled successfully.